changes to authentication

+++ lab/src/main/java/mk/ukim/finki/ib/lab/services/implementation/UserService.java
@@ -2,6 +2,7 @@ package mk.ukim.finki.ib.lab.services.implementation;
 
 import mk.ukim.finki.ib.lab.models.User;
 import mk.ukim.finki.ib.lab.models.enums.UserRole;
+import mk.ukim.finki.ib.lab.models.exceptions.AdminCantChangeAdminException;
 import mk.ukim.finki.ib.lab.repository.UserRepository;
 import mk.ukim.finki.ib.lab.services.interfaces.IUserService;
 import org.springframework.stereotype.Service;
@@ -22,7 +23,12 @@ public class UserService implements IUserService {
     }
 
     @Override
-    public void deleteById(int id) {
+    public void deleteById(int id, User currentUser) throws AdminCantChangeAdminException {
+
+        if (!currentUser.getRole().equals(UserRole.SUPERADMIN)){
+            throw new AdminCantChangeAdminException();
+        }
+
         this.userRepository.deleteById(id);
     }
 
@@ -32,7 +38,12 @@ public class UserService implements IUserService {
     }
 
     @Override
-    public void changeRole(int id, UserRole role) {
+    public void changeRole(int id, UserRole role, User currentUser) throws AdminCantChangeAdminException {
+
+        if (currentUser.getRole().equals(UserRole.ADMIN) && role.equals(UserRole.ADMIN)){
+            throw new AdminCantChangeAdminException();
+        }
+
         User user = userRepository.findById(id);
 
         user.setRole(role);
